change decimalfields in models to IntegerFields

+++ Football_travel/models.py
@@ -8,9 +8,9 @@ from cloudinary.models import CloudinaryField
 class Trip(models.Model):
     trip_name = models.CharField(max_length=50)
     trip_id = models.AutoField(primary_key=True)
-    seats = models.DecimalField(decimal_places=0, max_digits=2)
-    remaining_seats = models.DecimalField(decimal_places=0, max_digits=2)
-    price = models.DecimalField(decimal_places=2, max_digits=6)
+    seats = models.IntegerField()
+    remaining_seats = models.IntegerField()
+    price = models.IntegerField()
     date = models.DateField()
 
     def __str__(self):
@@ -23,7 +23,7 @@ class Booking(models.Model):
     created_on = models.DateTimeField(auto_now_add=True)
     comment = models.TextField(blank=True)
     trip_booked = models.ForeignKey(Trip, on_delete=models.CASCADE)
-    seats_required = models.DecimalField(decimal_places=0, max_digits=2)
+    seats_required = models.IntegerField()
 
 
 def __str__(self):
